% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii-filter.R
\name{ascii_filter}
\alias{ascii_filter}
\title{ASCII filter}
\usage{
ascii_filter(nr, sc_factor = 128, intensity = 0.2)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{sc_factor}{A numeric scalar; the scale factor.}
}
\value{
A \code{nativeRaster} object.
}
\description{
ASCII filter
}
