% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deform.R
\name{deform}
\alias{deform}
\title{Deformation}
\usage{
deform(nr, freq = c(10, 10), amp = c(0.1, 0.1))
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{freq}{A numeric vector of length 2;
the frequency of the deformation.}

\item{amp}{A numeric vector of length 2;
the amplitude of the deformation.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Deforms the image by sampling it at
\code{uv + float2(sin(uv.y * freq.x) * amp.x, cos(uv.x * freq.y) * amp.y)}.
}
