% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godray.R
\name{godray}
\alias{godray}
\title{Godray effect}
\usage{
godray(
  nr,
  parallel = FALSE,
  angle = pi/6,
  center = c(0, 0),
  gain = 0.5,
  lacunarity = 2.5,
  alpha = 1,
  octaves = 5,
  time = 0
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{parallel}{A logical scalar; whether to use parallel lighting.}

\item{angle}{A numeric scalar; the angle of the light.
If \code{parallel} is \code{FALSE}, this is ignored.}

\item{center}{A numeric vector of length 2;
the offset coordinates to roughly change the position of the light.
If \code{parallel} is \code{TRUE}, this is ignored.}

\item{gain}{A numeric scalar; the gain of the effect.}

\item{lacunarity}{A numeric scalar; the lacunarity of the effect.}

\item{alpha}{A numeric scalar; the alpha of the effect.}

\item{octaves}{A numeric scalar; the octaves of turbulence.}

\item{time}{A numeric scalar; ellapsed time.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Godray effect
}
