% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeshadr-package.R
\name{nr_new}
\alias{nr_new}
\title{Create a \code{nativeRaster} image}
\usage{
nr_new(width, height, fill = "white")
}
\arguments{
\item{width, height}{Image dimensions in pixels}

\item{fill}{Background fill color as a character string. Either a standard R color
(e.g. 'blue', 'white')
or a hex color of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}
}
\value{
\code{nativeRaster}
}
\description{
A \code{nativeRaster} in R looks like an integer matrix, but is interpreted
differently by graphics devices:
}
\details{
\itemize{
  \item{The data should be treated as RGBA pixels in row-major ordering}
  \item{Each 32-bit integer should be interpreted as 4-bytes - one for each
        of the R, G, B and A color channels}
}
}
\keyword{internal}
