% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb-split.R
\name{rgb_split}
\alias{rgb_split}
\title{Chromatic aberration effect}
\usage{
rgb_split(nr, r = c(0.01, 0), g = c(0, 0.01), b = c(0, 0))
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{r, g, b}{A numeric vector of length 6;
X and Y offsets for the red, green and blue channels
(in uv coordinates).}
}
\value{
A \code{nativeRaster} object.
}
\description{
Chromatic aberration effect
}
