% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeshadr-package.R
\docType{package}
\name{nativeshadr-package}
\alias{nativeshadr}
\alias{nativeshadr-package}
\title{nativeshadr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
