% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeshadr-package.R
\docType{package}
\name{nativeshadr-package}
\alias{nativeshadr}
\alias{nativeshadr-package}
\title{nativeshadr: Introduction of 'HLSL' Syntax to 'Rcpp'}
\description{
Brings 'HLSL'-like syntax to 'Rcpp' codes via the 'HLSL++' library \url{https://github.com/redorav/hlslpp}.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Other contributors:
\itemize{
  \item Emilio López (Author of the 'HLSL++' library) [copyright holder]
}

}
\keyword{internal}
