% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro-filter.R
\name{retro_filter}
\alias{retro_filter}
\title{Retro look effect}
\usage{
retro_filter(nr, asp = rev(dim(nr)/100), distort = TRUE, time = 0)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{asp}{A numeric vector of length 2; factor to scale the offset of vignette effect.}

\item{distort}{A logical scalar; whether to apply barrel distortion.}

\item{time}{A numeric scalar; ellapsed time.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Retro look effect
}
