% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripple.R
\name{ripple}
\alias{ripple}
\title{Ripple effect}
\usage{
ripple(nr, center = dim(nr)/2, freq = 50, amp = 0.01)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{center}{A numeric vector of length 2.
The distortion of the effect increases the farther they are from this point.}

\item{freq}{A numeric scalar; the frequency of the effect.}

\item{amp}{A numeric scalar; the amplitude of the effect.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Adds a ripple effect to the image by sampling it at
\code{uv + sin(length(uv - center) * freq) * amp}.
}
