% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shockwave.R
\name{shockwave}
\alias{shockwave}
\title{Shockwave effect}
\usage{
shockwave(
  nr,
  center = rev(dim(nr))/2,
  speed = log(min(dim(nr))),
  amp = 50,
  wave_length = 160,
  brigthness = 1,
  radius = -1,
  time = 1
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{center}{A numeric vector of length 2;
the center coordinates where the shockwave starts.}

\item{speed}{A numeric scalar;
the speed about the shockwave ripples out (in pixels per \code{time}).}

\item{amp}{A numeric scalar;
the amplitude of the shockwave.}

\item{wave_length}{A numeric scalar;
the wave length of the shockwave.}

\item{brigthness}{A numeric scalar;
the brigthness of the shockwave.}

\item{radius}{A numeric scalar;
The maximum radius of the shockwave. Set to \code{-1} for infinite radius.}

\item{time}{A numeric scalar; ellapsed time.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Shockwave effect
}
