% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twist.R
\name{twist}
\alias{twist}
\title{Twist effect}
\usage{
twist(nr, offset = dim(nr)/2, angle = pi, radius = 100, padding = 20)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{offset}{A numeric vector of length 2;
the offset coordinates to change the position of the center of the effect.}

\item{angle}{A numeric scalar; the angle of the twist.}

\item{radius}{A numeric scalar; the radius of the twist.}

\item{padding}{A numeric scalar; padding for the filter area.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Twist effect
}
